test_gdal_rasterize.py: fix failure on test_gdal_rasterize_cleanup() if test_gdal_rasterize_7 is skipped
git-svn-id: https://svn.osgeo.org/gdal/trunk@29612 f0d54148-0727-0410-94bb-9a71ac55c965

+++ autotest/utilities/test_gdal_rasterize.py
@@ -409,7 +409,8 @@ def test_gdal_rasterize_cleanup():
     os.unlink('tmp/test_gdal_rasterize_6.prj')
 
     gdal.GetDriverByName('GTiff').Delete( 'tmp/test_gdal_rasterize_7.tif' )
-    os.unlink('tmp/test_gdal_rasterize_7.csv')
+    if os.path.exists('tmp/test_gdal_rasterize_7.csv'):
+        os.unlink('tmp/test_gdal_rasterize_7.csv')
 
     return 'success'
